Re-Sync from older version
Re-Sync from older version

+++ cron/gpio_ds18b20.py
@@ -70,39 +70,18 @@ def insertDB(IDs, temperature):
 			cur.execute('INSERT INTO messages_in(`sync`, `purge`, `node_id`, `child_id`, `sub_type`, `payload`, `datetime`) VALUES(%s,%s,%s,%s,%s,%s,%s)', (0, 0, IDs[i], 0, 0, round(temperature[i],2), time.strftime("%Y-%m-%d %H:%M:%S")))
 			con.commit()
 			#Check is sensor is attached to a zone which is being graphed
-			cur.execute('SELECT temperature_sensors.id, temperature_sensors.zone_id, nodes.node_id, temperature_sensors.sensor_child_id, temperature_sensors.name, temperature_sensors.graph_it FROM temperature_sensors, `nodes` WHERE (temperature_sensors.sensor_id = nodes.`id`) AND  nodes.node_id = "' + IDs[i] + '" AND temperature_sensors.sensor_child_id = 0 AND temperature_sensors.graph_it = 1 LIMIT 1;')
+			cur.execute('SELECT * FROM `zone_view` where sensors_id = (%s) LIMIT 1;', [IDs[i]])
 			results =cur.fetchone()
 			if cur.rowcount > 0:
-				sensor_to_index = dict(
-				(d[0], i)
-				for i, d
-				in enumerate(cur.description)
-				)
-				sensor_id = int(results[sensor_to_index['id']])
-				sensor_name = results[sensor_to_index['name']]
-				zone_id = int(results[sensor_to_index['zone_id']])
-				graph_it = int(results[sensor_to_index['graph_it']])
-				if graph_it == 1:
+				zone_id = int(results[3])
+				name = results[5]
+				type = results[6]
+				category = int(results[7])
+				graph_it = int(results[8])
+				if category < 2 and graph_it == 1:
 					print(bc.dtm + time.ctime() + bc.ENDC + ' - Adding Temperature Reading to Graph Table From Node ID:', IDs[i], ' PayLoad:', temperature[i])
-					if zone_id == 0:
-						cur.execute('INSERT INTO zone_graphs(`sync`, `purge`, `zone_id`, `name`, `type`, `category`, `node_id`,`child_id`, `sub_type`, `payload`, `datetime`) VALUES(%s,%s,%s,%s,%s,%s,%s,%s,%s,%s,%s)', (0,0,sensor_id,sensor_name,"Sensor",0,IDs[i],0,0, round(temperature[i],2),time.strftime("%Y-%m-%d %H:%M:%S")))
-						con.commit()
-					else:
-						cur.execute('SELECT * FROM `zone_view` where id = (%s) LIMIT 1;', (zone_id, ))
-						results =cur.fetchone()
-						if cur.rowcount > 0:
-							zone_view_to_index = dict(
-							(d[0], i)
-							for i, d
-							in enumerate(cur.description)
-							)
-							zone_name = results[zone_view_to_index['name']]
-							type = results[zone_view_to_index['type']]
-							category = int(results[zone_view_to_index['category']])
-							if category != 2:
-								#cur.execute('INSERT INTO zone_graphs(`sync`, `purge`, `zone_id`, `name`, `type`, `category`, `node_id`,`child_id`, `sub_type`, `payload`, `datetime`) VALUES(%s,%s,%s,%s,%s,%s,%s,%s,%s,%s,%s)', (0,0,sensor_id,zone_name,type,category,[str(IDs[i])],0,0, round(temperature[i],2),time.strftime("%Y-%m-%d %H:%M:%S")))
-								cur.execute('INSERT INTO zone_graphs(`sync`, `purge`, `zone_id`, `name`, `type`, `category`, `node_id`,`child_id`, `sub_type`, `payload`, `datetime`) VALUES(%s,%s,%s,%s,%s,%s,%s,%s,%s,%s,%s)', (0,0,zone_id,zone_name,type,category,IDs[i],0,0,round(temperature[i],2),time.strftime("%Y-%m-%d %H:%M:%S")))
-								con.commit()
+					cur.execute('INSERT INTO zone_graphs(`sync`, `purge`, `zone_id`, `name`, `type`, `category`, `node_id`,`child_id`, `sub_type`, `payload`, `datetime`) VALUES(%s,%s,%s,%s,%s,%s,%s,%s,%s,%s,%s)', (0,0,zone_id,name,type,category,IDs[i],0,0,round(temperature[i],2),time.strftime("%Y-%m-%d %H:%M:%S")))
+					con.commit()
 					cur.execute('DELETE FROM zone_graphs WHERE node_id = (%s) AND datetime < CURRENT_TIMESTAMP - INTERVAL 24 HOUR;', [IDs[i]])
 					con.commit()
 		con.close()
